servers/genius/pyproject.toml
[tool.poetry]
name = "genius"
version = "0.0.0"
description = ""
authors = ["JY"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
django = "^5.0.1"
mysqlclient = "^2.2.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

